Delete README.md
This  game is built with HTML., CSS, and JavaScript.
In this game, you are running away from you chores! The player is to open each, starting from the first door, which possibly leads to vacation opportunity. However, if you encounter the chore robot, the game is over and you are inevitably faced with doing your chores! So hopefully you don't encounter the robot.